(maint) - remove Binford2k module
