+++ lib/puppet/functions/node_decrypt.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../puppet_x/binford2k/node_encrypt'
+require_relative '../../puppet_x/node_encrypt'
 
 # Decrypt data with node_encrypt. This is intended to be used as a
 # Deferred function on the _agent_ via the node_encrypted::secret wrapper.
@@ -12,7 +12,7 @@ Puppet::Functions.create_function(:node_decrypt) do
 
   def decrypt(content)
     Puppet::Pops::Types::PSensitiveType::Sensitive.new(
-      PuppetX::Binford2k::NodeEncrypt.decrypt(content),
+      PuppetX::NodeEncrypt.decrypt(content),
     )
   end
 end
