+++ lib/puppet/functions/node_encrypt.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../puppet_x/binford2k/node_encrypt'
+require_relative '../../puppet_x/node_encrypt'
 
 # @summary
 #   Encrypt data with node_encrypt.
@@ -16,7 +16,7 @@ Puppet::Functions.create_function(:node_encrypt) do
 
   def simple_encrypt(content)
     certname = closure_scope['clientcert']
-    PuppetX::Binford2k::NodeEncrypt.encrypt(content, certname)
+    PuppetX::NodeEncrypt.encrypt(content, certname)
   end
 
   def sensitive_encrypt(content)
