+++ spec/functions/node_encrypt_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-require 'puppet_x/binford2k/node_encrypt'
+require 'puppet_x/node_encrypt'
 
 describe 'node_encrypt' do
   let(:node) { 'testhost.example.com' }
@@ -11,13 +11,13 @@ describe 'node_encrypt' do
   end
 
   it 'receives foobar and returns encrypted' do
-    expect(PuppetX::Binford2k::NodeEncrypt).to receive(:encrypt).with('foobar', 'testhost.example.com').and_return('encrypted')
+    expect(PuppetX::NodeEncrypt).to receive(:encrypt).with('foobar', 'testhost.example.com').and_return('encrypted')
     expect(scope.function_node_encrypt(['foobar'])).to eq('encrypted')
   end
 
   if defined?(Puppet::Pops::Types::PSensitiveType::Sensitive)
     it 'receives sensitive value and returns encrypted' do
-      expect(PuppetX::Binford2k::NodeEncrypt).to receive(:encrypt).with('foobar', 'testhost.example.com').and_return('encrypted')
+      expect(PuppetX::NodeEncrypt).to receive(:encrypt).with('foobar', 'testhost.example.com').and_return('encrypted')
       expect(scope.function_node_encrypt([Puppet::Pops::Types::PSensitiveType::Sensitive.new('foobar')])).to eq('encrypted')
     end
   end
